[Tools.ServiceReference] * merge Tools.ServiceReference.py and ServiceReference.py
