--- lib/python/Tools/ServiceReference.py
@@ -1,50 +1,0 @@
-from enigma import eServiceReference, eServiceReferenceDVB
-
-# type 1 = digital television service
-# type 4 = nvod reference service (NYI)
-# type 17 = MPEG-2 HD digital television service
-# type 22 = advanced codec SD digital television
-# type 24 = advanced codec SD NVOD reference service (NYI)
-# type 25 = advanced codec HD digital television
-# type 27 = advanced codec HD NVOD reference service (NYI)
-# type 2 = digital radio sound service
-# type 10 = advanced codec digital radio sound service
-# type 31 = High Efficiency Video Coding digital television
-# type 32 = High Efficiency Video Coding digital television
-
-# Generate an eServiceRef query path containing
-# '(type == serviceTypes[0]) || (type == serviceTypes[1]) || ...'
-
-
-def makeServiceQueryStr(serviceTypes):
-	return ' || '.join(['(type == %d)' % x for x in serviceTypes])
-
-
-def serviceRefAppendPath(sref, path):
-	nsref = eServiceReference(sref)
-	nsref.setPath(nsref.getPath() + path)
-	return nsref
-
-
-service_types_tv_ref = eServiceReference(eServiceReference.idDVB, eServiceReference.flagDirectory, eServiceReferenceDVB.dTv)
-
-service_types_tv_ref.setPath(makeServiceQueryStr((
-	eServiceReferenceDVB.dTv,
-	eServiceReferenceDVB.mpeg2HdTv,
-	eServiceReferenceDVB.avcSdTv,
-	eServiceReferenceDVB.avcHdTv,
-	eServiceReferenceDVB.nvecTv,
-	eServiceReferenceDVB.nvecTv20,
-	eServiceReferenceDVB.user134,
-	eServiceReferenceDVB.user195,
-)))
-
-service_types_radio_ref = eServiceReference(eServiceReference.idDVB, eServiceReference.flagDirectory, eServiceReferenceDVB.dRadio)
-service_types_radio_ref.setPath(makeServiceQueryStr((
-	eServiceReferenceDVB.dRadio,
-	eServiceReferenceDVB.dRadioAvc,
-)))
-
-
-def hdmiInServiceRef():
-	return eServiceReference(eServiceReference.idServiceHDMIIn, eServiceReference.noFlags, eServiceReferenceDVB.dTv)
